Splitting FieldTypeContainer into its own file
